==> src/field_chemistry/01_chem_prep.qmd <==
---
title: "chemistry prep"
format: html
  self-contained: true
code-fold: true
warning: false
editor: visual
---

# Cameron Peak Fire Reservoir Water Analysis

This will be the primary document for cleaning data

#### Set Up

```{r}
source("package_loader.R")
source("field_chemistry/00_colors_and_groups.R")

```

### [**Goals:**]{.underline}

#### This Document will:

-   Clean reservoir data from Tim/Chuck and make it usable for our analysis

This looks like:

-   Making dates compatible

-   Adding location characteristics

-   Graph changes in key nutrients over time

    -   Focus on NO3, PO4,

## Cleaning data

-   Add Location data to chemistry
-   Adding site_code
-   Joining data from above
-   Add year column
-   Creating Seperate DF for Mainstem Work

```{r}


reservoir_chemistry <- read_csv('data/field_chemistry/raw/ReservoirChemistry_073123.csv')%>%
  mutate(Date = parse_date_time(Date, 
                                  orders = c('%m/%d/%y',
                                             '%d-%b-%y')))%>%
  filter(SampleType == "NORM") %>%
  mutate (FCW_Number = as.numeric(gsub('FCW ', '', IDNo)), 
          Year =  year(Date), 
          site_code = SiteLabel)%>%
  select(-c(SiteLabel, IDNo, SiteDescr, SampleType))%>%
  left_join(select(Sites,c(site_code,Lat,Long, location, Campaign, watershed, status)),by="site_code") %>%
   left_join(select(buffer_sbs,c(watershed,Buffer_Level)),by="watershed") %>%
  left_join(select(watershed_sbs,c(site_code,Watershed_Level)),by="site_code")%>%
  left_join(select(dist_from_pbd,c(site_code,distance_upstream_km)),by="site_code") %>%
   mutate(season=case_when(between(Date, as.Date("2022-01-01"), as.Date("2022-06-09")) ~ "SPRING", 
                           between(Date, as.Date("2021-01-01"), as.Date("2021-06-09")) ~ "SPRING",
                           between(Date, as.Date("2022-06-10"), as.Date("2022-09-24")) ~ "SUMMER", 
                           between(Date, as.Date("2021-06-10"), as.Date("2021-09-24")) ~ "SUMMER",
                          .default =  "FALL")) %>%
  mutate(Distance = case_when(site_code == 'JOEI' ~ '1 - JOEI',
                              site_code == 'JOER' ~ '2 - JOER',
                              site_code == 'CBRI' ~ '3 - CBRI',
                              site_code == 'CBRR' ~ '4 - CBRR',
                              site_code == 'CHD' ~ '5 - CHD',
                              site_code == 'BRNR' ~ '6',
                              site_code == 'BMD' ~ '7',
                              site_code == 'LNGR' ~ '8',
                              site_code == 'LNGO' ~ '9',
                              site_code == 'PTRR' ~ '99',
                              site_code == 'PTRO' ~ '999',
                              site_code == 'COMI' ~ '11 - COMI',
                              site_code == 'COMR' ~ '12 - COMR',
                              site_code == 'COMO' ~ '13 - COMO',
                              site_code == 'HORI' ~ '14 - HORI',
                              site_code == 'HORR' ~ '15 - HORR',
                              site_code == 'HORO' ~ '16 - HORO',
                              site_code == 'BEAV' ~ '17 - BEAV',
                              site_code == 'JWC' ~ '18 - JWC',
                              site_code == 'PJW' ~ '19 - PJW',
                              site_code == 'SLEP' ~ '20 - SLEP',
                              site_code == 'PBR' ~ '21 - PBR',
                              site_code == 'SFM' ~ '22 - SFM',
                              site_code == 'PSF' ~ '23 - PSF',
                              site_code == 'PNF' ~ '24 - PNF',
                              site_code == 'PBD' ~ '25 - PBD'
                              )) %>%
  mutate(Xdistance = case_when(site_code == 'JOEI' ~ 1,
                              site_code == 'JOER' ~ 2,
                              site_code == 'CBRI' ~ 3,
                              site_code == 'CBRR' ~ 4,
                              site_code == 'CHD' ~ 5,
                              site_code == 'LNGR' ~ 6,
                              site_code == 'BRNR' ~ 7,
                              site_code == 'COMI'~ 12,
                              site_code == 'COMR'~ 13,
                              site_code == 'COMO'~ 14,
                              site_code == 'HORI'~ 15,
                              site_code == 'HORR'~ 16,
                              site_code == 'HORO'~ 17,
                              site_code == 'BEAV'~ 18,
                              site_code == 'JWC'~ 19,
                              site_code == 'PJW'~ 20,
                              site_code == 'SLEP'~ 21,
                              site_code == 'PBR'~ 22,
                              site_code == 'SFM'~ 23,
                              site_code == 'PSF'~ 24,
                              site_code == 'PNF'~ 25,
                              site_code == 'PBD'~ 26)) %>%
  dplyr::arrange(as.factor(status)) %>%
  dplyr::filter(!watershed %in% c('Hourglass','Comanche')) %>%
  mutate(order = case_when(site_code =="JOEI" ~ "03",
                        site_code == "LNGR" ~ "01",
                        site_code == "LNGO" ~ "02",
                        site_code == "JOER" ~ "04",
                        site_code == "CBRI" ~ "05",
                        site_code == "CBRR" ~ "06",
                        site_code == "CHD" ~ "07",
                        site_code == "PTRR" ~ "08",
                        site_code == "PTRO" ~ "09",
                        site_code == "BRNR" ~ "10",
                        site_code == "BMD" ~ "11",
                        site_code == "COMI"~ "12",
                        site_code == "COMR" ~ "13",
                        site_code == "COMO"~ "14",
                        site_code == "HORI"~ "15",
                        site_code == "HORR"~ "16",
                        site_code == "HORO"~ "17",
                        site_code == "BEAV"~ "18",
                        site_code == "JWC"~ "19",
                        site_code == "PJW"~ "20",
                        site_code == "SLEP"~ "21",
                        site_code == "PBR"~ "22",
                        site_code == "SFM"~ "23",
                        site_code == "PSF" ~ "24",
                        site_code == "PNF"~ "25",
                        site_code == "PBD"~ "26",))%>%
  dplyr::arrange(as.factor(status))

# write a feather file named reservoir_chem_2021_2022.feather
write_feather(reservoir_chemistry, "data/field_chemistry/cleaned/reservoir_chem_2021_2022_062023.feather")


```

## Data Checks

```{r}



# number of samples per site per year
res_campaign_by_year <- reservoir_chemistry%>%
  group_by(Year, Campaign)%>%
  summarise( sample_count = n())%>%
  ungroup()%>%
  ##pivot the data so that each campaign is a column
  pivot_wider(names_from = "Campaign", values_from = "sample_count")%>%
  # replace all NAs with 0
  mutate_at(vars(-Year), ~replace(., is.na(.), 0))

print(res_campaign_by_year, n = length(res_campaign_by_year$Year))


#Determning number of Samples....

rmrs_chemistry <- read.csv("data/field_chemistry/raw/CamPkChem.csv")%>%
  select(Date, SiteLabel ,ChlA)%>%
  mutate(Date = as.Date.character(Date,format = "%d-%b-%y"),
         year = year(Date),
    campaign = case_when(SiteLabel %in% all_res_system ~ "Reservoir", 
                         grepl(pattern = "ISCO", SiteLabel) ~"ISCO",
                         SiteLabel %in% mainstem_sites ~"Mainstem", 
                         SiteLabel %in% genomic_sites ~ "In-Town Genomic", 
                         SiteLabel == "BEAV" ~ "Reservoir", 
                         .default = "Tributary"))

  
#samples by year and campaign

campaign_by_year <- rmrs_chemistry%>%
  group_by(year, campaign)%>%
  summarise( sample_count = n())%>%
  ungroup()%>%
  ##pivot the data so that each campaign is a column
  pivot_wider(names_from = "campaign", values_from = "sample_count")%>%
  # replace all NAs with 0
  mutate_at(vars(-year), ~replace(., is.na(.), 0))




print(campaign_by_year, n = length(campaign_by_year$year)) 

#create a similar data frame as campaign_by_year but using the reservoir_chemistry data frame


# number of unique chla measurements
unique_chla <- rmrs_chemistry%>%
  filter(!is.na(ChlA))%>%
  group_by(year, campaign)%>%
  dplyr::summarise( chla_samples = n())%>%
  ungroup()
print(unique_chla)


```

### Importing ICP chem

```{r}

res_icp <- read_csv("data/field_chemistry/raw/ReservoirStudy_ICPData.csv")%>%
  mutate(Date = as.Date(Date, format = "%d-%b-%y"), 
         Al = as.numeric(gsub("x", "", Al_ppb)), 
         P = as.numeric(gsub("x", "", P_ppm)), 
         Mn = as.numeric(gsub("x", "", Mn_ppb)),
         Fe = as.numeric(gsub("x", "", Fe_ppb)))%>%
  select(FCW_Number = EraSamp, 
         Al, P, Mn, Fe)


res_longer <- reservoir_chemistry%>%
  left_join(res_icp, by = "FCW_Number")%>%
    select(-c(Lat, Long, order, Xdistance, Distance, Site))%>%
  pivot_longer(cols = c(ChlA, Turbidity, TSS, DOC, DTN, pH,ANC, Na, NH4, K, Mg, Ca, F, Cl, PO4, SO4, SC,NO3, Al, P, Mn, Fe), names_to = "Parameter",values_to = "Measurement" )


write_feather(res_longer, sink = "data/field_chemistry/cleaned/reservoir_chemistry_longer.feather")

  
```

## Exporting data to csv for Poudre Portal

```{r writing}
write.csv(reservoir_chemistry, 
          "data/field_chemistry/Poudre_Portal_Data/prepped_reservoir_chemistry.csv")

write.csv(filter(reservoir_chemistry, location == "Mainstem"),
          "data/field_chemistry/Poudre_Portal_Data/prepped_mainstem_chemistry.csv")

reservoir_chemistry_concise <- reservoir_chemistry%>%
  select(Site, site_code, Date,Lat,Long,watershed,Watershed_Level, status, Turbidity,TSS,ChlA,pH,ANC,SC,Na,NH4,K,Mg,Ca,F,Cl,NO3,PO4,SO4, DOC, DTN)

write_csv(reservoir_chemistry_concise, 
          "data/field_chemistry/Poudre_Portal_Data/CPF_reservoir_chemistry_up_to_071023.csv")

```

## Combining RMRS and FcGov datasets

```{r}



#importing 
fcgov_mainstem <- read.csv("data/field_chemistry/raw/fcgov_mainstem_chem.csv", na.strings = '*')%>%
  mutate(#Date = as_date(Date, format = "%m/%d/%Y"),
         Date = make_date(Year, Month, Day),
         site_code = ifelse(ShortDesc %in% c("BMR", "BMD"), "BMD", 
                     ifelse(ShortDesc %in% c("CBR", "CHD"), "CHD",ShortDesc)))%>%
 dplyr::rename(NO3 = NO3_N, 
        K = K_S,
         Na = Na_S, 
         SC = SpCon, 
        PO4 = oPhos, 
        Mg = Mg_S, 
        Ca = Ca_S)%>%
  mutate(status = ifelse(site_code %in% fcgov_study, "Burned", "Unburned"),
         dayofyear = yday(Date), 
         collector = "City of Fort Collins")%>%
  dplyr::select(ShortDesc, site_code,Year, Date ,Turbidity, pH, SC, Na, K, Mg, Ca, Cl, NO3, PO4, SO4, DO, Temp,collector, status )

write_feather(fcgov_mainstem, sink = "data/field_chemistry/cleaned/fcgov_mainstem")


#combining with RMRS data
combined_mainstem <- filter(reservoir_chemistry, location == "Mainstem")%>%
  mutate(collector = "RMRS")%>%
  bind_rows(fcgov_mainstem)%>%
  mutate(dayofyear = yday(Date), 
         Year = year(Date), 
         fire = ifelse(Date > "2020/8/1", "Cameron Peak", "High Park"))%>%
  select(-c(Lat, Long, location, Campaign, watershed, status, season))
  combined_mainstem$site_code = factor(combined_mainstem$site_code, levels = c("JWC", "PJW","SLEP", "PBR", "SFM","PSF","PNF", "PBD"  ))

```

### Graphing

```{r}

combined_graph <- filter(combined_mainstem, site_code %in% c("JWC" , "PJW", "PBR", "PSF", "PNF", "PBD")
                                                              & Year <= 2021 & Year >= 2019)%>%
  ggplot()+
  geom_point(aes(x = dayofyear, y = NO3, group = interaction(Year, collector), color = Year, shape = collector),size = 3, alpha = .5)+
  geom_line(aes(x = dayofyear, y = NO3, group = interaction(Year, collector),  color = Year), size = 1)+
  scale_shape_manual(values =  c(15, 17))+
  theme_bw()+
  facet_wrap(~site_code, ncol = 3)+
  scale_color_viridis_b()

#ggplotly(combined_graph)



#graphing yearly changes in nutrients pre/post fire
combined_graph <- filter(combined_mainstem, site_code %in% c("PBR"))%>%
  ggplot()+
  geom_point(aes(x = dayofyear, y = NO3, color = Year, shape = collector),size = 3, alpha = .5)+
  geom_line(aes(x = dayofyear, y = NO3, group = Year,  color = Year), size = 1)+
  scale_shape_manual(values =  c(15, 17))+
  theme_bw()+
   scale_color_viridis_c()

plot(combined_graph)

#ggsave("output/mainstem_timeline_no3_2008_2022_FCGOV_RMRS.jpg", width = 12, height = 9, dpi = 300, units = "in" )

combined_boxplot_no3 <- filter(combined_mainstem, site_code %in%c("PBR" ,"PJW", "JWC","PBD"))%>%
  ggplot(aes(x= site_code , y=NO3, fill =site_code )) + 
geom_boxplot()+
   geom_dotplot( stackgroups = FALSE, binaxis = "y", method = "histodot",stackdir = "center", position = "dodge", binwidth = .05, dotsize = 1, alpha = .7)+
  stat_summary(fun.y = median, fun.ymin = median, fun.ymax = median,
                 geom = "crossbar", width = .75)+
  theme_bw(base_size = 20)+
  facet_wrap(~fire+collector)
plot(combined_boxplot_no3)

#ggsave("output/mainstem_boxplot_no3_2008_2022_FCGOV_RMRS.jpg", width = 14, height = 9, dpi = 300, units = "in" )

combined_boxplot_K <- filter(combined_mainstem, site_code %in%c("PBR" ,"PJW", "JWC","PBD"))%>%
  ggplot()+
  geom_boxplot(aes(x = site_code, y = K, middle = mean(K), group = site_code, fill = site_code))+
  theme_bw(base_size = 20)+
  facet_wrap(~fire+collector)
plot(combined_boxplot_K)

combined_boxplot_PO4 <- filter(combined_mainstem, site_code %in%c("PBR" ,"PJW", "JWC","PBD"))%>%
  ggplot()+
  geom_boxplot(aes(x = site_code, y = PO4, middle = mean(PO4), group = site_code, fill = site_code))+
  theme_bw(base_size = 20)+
  facet_wrap(~fire+collector)
plot(combined_boxplot_PO4)

combined_boxplot_SC <- filter(combined_mainstem, site_code %in%c("PBR" ,"PJW", "JWC","PBD"))%>%
  ggplot()+
  geom_boxplot(aes(x = site_code, y = SC, middle = mean(SC), group = site_code, fill = site_code))+
  theme_bw(base_size = 20)+
  facet_wrap(~fire+collector)
plot(combined_boxplot_SC)
```
