==> src/field_chemistry/02_chla_graphing.qmd <==
---
title: "chla_graphing"
format: html
editor: visual
self-contained: true
---

## Goals

This document is to work on graphs related to nutrients from Reservoirs/ mainstem

### Importing

Loading packages and creating lists of sites based on location/grouping

```{r}
source("package_loader.R")
source("field_chemistry/00_colors_and_groups.R")


reservoir_chemistry <- read_feather("data/field_chemistry/cleaned/reservoir_chem_2021_2022_062023.feather")%>%
  mutate(location = factor(location, levels = c("Inflow","Reservoir", "Outflow", "Stream", "Mainstem")))


```

# Chl-A Plots

## Goals:

-   Replicate Katie's Plots from last year for this year's data

-   Facet the 2 years to compare them to each other

-   New graph ideas?

## Dataframe for calculations

```{r}

chla_status_site_year <- reservoir_chemistry%>%
  filter(!is.na(ChlA))%>%
  mutate(eu_status  = case_when(ChlA >= 2.6 &ChlA <= 8 ~ "Mesotrophic", 
                                ChlA > 8 ~ "Eutrophic",
                                ChlA < 2.6 ~ "Oligotrophic"))%>%
  group_by(site_code, Year, eu_status)%>%
  summarise(eu_stats = n())%>%
  ungroup()%>%
  pivot_wider(names_from = eu_status, values_from = eu_stats)%>%
  mutate_at(vars(-Year), ~replace(., is.na(.), 0))%>%
  rowwise()%>%
  mutate(total_samples = sum(c(Eutrophic, Mesotrophic, Oligotrophic)),
    perc_eu = (Eutrophic/total_samples)*100, 
         perc_meso = (Mesotrophic/total_samples)*100,
         perc_meso_eu = ((Eutrophic+Mesotrophic)/total_samples)*100, 
        perc_oligo = (Oligotrophic/total_samples) *100,
    Year = as.character(Year))


long_chla_status <- chla_status_site_year%>%
  select(site_code, Year, total_samples, perc_eu, perc_meso, perc_meso_eu, perc_oligo)%>%
  pivot_longer(cols = c(perc_eu, perc_meso, perc_meso_eu, perc_oligo), names_to = "category", values_to = "value" )

status_plot <- long_chla_status%>%
  filter(site_code %in% mainstem_Res_only )%>%
  ggplot(aes(x = Year, y = value,group = site_code, color =  site_code))+
  scale_color_manual(values = colorsBS_site_code)+
 geom_line()+
   geom_point()+
  theme_bw(base_size = 15)+
facet_wrap(~category)
plot(status_plot)



chla_per_year <-reservoir_chemistry%>%
  select(ChlA, Year, site_code, Date, watershed)%>%
  na.omit()%>%
  group_by(Year)%>%
  summarise(count_year = n())


chla_per_site_year <-reservoir_chemistry%>%
  select(ChlA, Year, site_code, Date)%>%
  na.omit()%>%
  group_by(Year)%>%
  mutate(sites = length(unique(site_code)))
```

## Longitudinal Chl-A and NO3

-   Longitudinal chla for 3 subsets

    -   Main study reservoirs

    -   SF reservoir network

    -   Mainstem

### Mainstem Reservoirs

Longitudinal view of NO3 and chla from JOEI to JOER

includes a highlighted graph for presentations

```{r}
  
  choosen_dates_chambers <- as.Date(c(#"2022-06-01",
            "2022-06-14",
            #"2022-07-05",
            "2022-07-26",
            "2022-08-16",
                #"2022-08-23",  
                #"2022-08-30",
                  #"2022-09-06",
              #"2022-09-13",
                  # "2022-09-20",
                  "2022-09-27",
                  #"2022-10-04",
                  #"2022-10-11",
                  "2022-10-18"
                  #"2022-10-25",
                  #"2022-11-01"
  ))
  
  
  
  #chambers dataframe
  chambers_longitudinal <- reservoir_chemistry %>%
    dplyr::filter(!is.na(ChlA)) %>%
    dplyr::filter(site_code %in% c("JOEI", "JOER", "CBRR", "CBRI", "CHD")) %>%
    dplyr::filter(Date %in% choosen_dates_chambers)%>%
    mutate(Date=as.character(Date),
           location_type = ifelse(location == "Outflow"|location =="Inflow", "Stream", "Reservoir"))
  
  ###PLOTING PART OF FUNCTION
  
  #chla graph
  chambers_chla_long <- chambers_longitudinal%>%
    ggplot(aes(x=Distance, y=ChlA, color=Date)) +
    geom_point( aes(shape = location_type), size=8, alpha = .7) +
    geom_line(aes(x=Xdistance), size=2, alpha = .7) +
    geom_point(aes(shape = location_type), size=8, alpha = .7) +
    scale_color_manual(values = season_color_vals)+
    scale_fill_manual(values = season_color_vals)+
    theme_bw(base_size=30) +
    labs(shape = "", color="") +
    theme(axis.text.x = element_blank(), 
          legend.text = element_text(size = 18)) +
    xlab("") +
    ylab("Chlorophyll a (μg/L)")
  
  #NO3 GRAPH
  chambers_no3_long <- chambers_longitudinal%>%
    ggplot(aes(x=Distance, y=NO3, color=Date)) +
    geom_point( aes(shape = location_type), size=8, alpha = .7) +
    geom_line(aes(x=Xdistance), size=2, alpha = .7) +
    geom_point(aes(shape = location_type), size=8, alpha = .7) +
    scale_color_manual(values = season_color_vals)+
    scale_fill_manual(values = season_color_vals) +
    theme_bw(base_size=30) +
    theme(legend.text = element_text(size =18))+
    labs(shape = "", color="") +
    scale_x_discrete(labels=c("1 - JOEI" = "Above Joe Wright", "2 - JOER" = "", "3 - CBRI" = "Above Chambers",
                              "4 - CBRR" = "", "5 - CHD" = "Below Chambers")) +
    xlab("Longitudinal Profile") +
    ylab("NO3 (mg/L)")
  
  
 chambers_plot <- ggarrange(chambers_chla_long,chambers_no3_long, ncol=1, nrow=2, common.legend=T)
  
chambers_plot
 #ggsave('output/2022_joer_cbrr_chla_no3_longitudinal_FINAL.jpg', width=15.5, height=10, dpi = 300)
 
 #highlighted version of graph for presentations
 
 chla_long_highlighted <- chambers_chla_long + gghighlight(Date == "2022-07-26", keep_scales = T)
 no3_long_highlighted <-chambers_no3_long + gghighlight(Date == "2022-07-26", keep_scales = T)
 
 
 ggarrange(chla_long_highlighted,no3_long_highlighted, ncol=1, nrow=2, common.legend=T)
 
 
 #ggsave('output/2022_joer_cbrr_chla_no3_longitudinal_highlighted.jpg', width=15, height=10)
```

### South Fork Reservoir

This is an overview of the SF Reservoir network

-   chla and no3 longitudinal

```{r}

 choosen_dates_southfork <- as.Date(c("2021-07-09",
                #"2021-09-21",
                "2022-06-29",
                "2022-07-18",
                "2022-08-31",
                "2022-09-28"
 ))
  #southfork dataframe
  SF_longitudinal <- reservoir_chemistry %>%
    dplyr::filter(!is.na(ChlA)) %>%
    dplyr::filter(site_code %in% c("COMI", "COMR", "COMO", "HORI", "HORR", "HORO", "BEAV", "SFM")) %>%
    dplyr::filter(Date %in% choosen_dates_southfork) %>%
    mutate(Date=as.character(Date), 
           location_type = ifelse(location == "Reservoir", "Reservoir", "Stream"))
  
  ##SOUTH FORK PLOTS
  
  SF_chla_long <- SF_longitudinal%>%
    ggplot(aes( x= Distance, y = ChlA, color = Date)) +
    geom_point(aes(shape = location_type), size=5) +
    geom_line(aes(group = Date), size=1) +
    geom_point(aes(shape = location_type), size=5) +
    scale_color_manual(values=season_color_vals) +
    scale_fill_manual(values=season_color_vals) +
    theme_bw(base_size=24) +
    labs(shape = "", color="") +
    theme(axis.text.x = element_blank(),
          legend.position = "top") +
    xlab("") +
    ylab("Chlorophyll a (μg/L)")
  
  #NO3
  
  SF_no3_long <- SF_longitudinal%>%
    ggplot(aes(x=Distance, y=NO3, color=Date)) +
    geom_point(aes(shape = location_type), size=5) +
    geom_line(aes(group = Date), size=1) +
    geom_point(aes(shape = location_type), size=5) +
    scale_color_manual( values = season_color_vals)+
    scale_fill_manual( values = season_color_vals) +
    theme_bw(base_size=24) +
    #theme(axis.text.x = element_blank()) +
    labs(shape = "", color="") +
    scale_x_discrete(labels=c("11 - COMI" = "COMI", "12 - COMR" = "COMR", "13 - COMO" = "COMO","14 - HORI" = "HORI", "15 - HORR" = "HORR", "16 - HORO" = "HORO", "17 - BEAV" = "BEAV", "22 - SFM" = "SFM")) +
    xlab("Longitudinal Profile") +
    ylab("NO3 (mg/L)")
  
  southfork_plot <- ggarrange(SF_chla_long,SF_no3_long, ncol=1, nrow=2, common.legend=T)
 southfork_plot 
 
 #ggsave('output/south_fork_reservoirs_chla_no3_longitudinal.jpg', width=15, height=15)


```

### Mainstem Longitudinal

```{r}

#chla

choosen_dates_mainstem <- as.Date(c(
  #"2022-05-16",
            "2022-05-31",
            #    "2022-06-07",
                "2022-07-11",
                #"2022-07-25",
                #"2022-08-08",
               # "2022-08-26",
                 #"2022-09-09",
                "2022-09-23",
                 #"2022-10-07",
                  "2022-10-21",
                  "2022-11-08"
))
#mainstem dataframe
mainstem_longitudinal <- reservoir_chemistry %>%
  dplyr::filter(!is.na(ChlA)) %>%
  dplyr::filter(site_code %in% c("JWC","PJW","SLEP","PBR","SFM","PSF","PNF","PBD"))%>%
  dplyr::filter(Date %in% choosen_dates_mainstem) %>%
  mutate(Date=as.character(Date))

##MAINSTEM PLOT

mainstem_chla_long <- mainstem_longitudinal%>%
  ggplot(aes(x= Distance, y= ChlA, color = factor(Date))) +
  geom_point( size=5) +
  geom_line(aes(group = Date), size=1) +
  geom_point( size=5) +
  scale_color_manual(values = season_color_vals)+
  scale_fill_manual(values = season_color_vals) +
  theme_bw(base_size=24) +
  labs(shape = "", color="") +
  xlab("") +
  ylab("Chlorophyll a (μg/L)")+
  #theme(axis.text.x = element_blank())
  #scale_color_manual(values = cbbPalette)
  scale_x_discrete(labels=c("18 - JWC" = "JWC", "19 - PJW" = "PJW", "20 - SLEP" = "SLEP","21 - PBR" = "PBR", "22 - SFM" = "SFM", "23 - PSF" = "PSF", "24 - PNF" = "PNF", "25 - PBD" = "Canyon Mouth" )) 



#no3
mainstem_no3_long <- mainstem_longitudinal%>%
  ggplot(aes(x=Distance, y=NO3, color=factor(Date))) +
  geom_point( size=5) +
  geom_line(aes(group = Date), size=1) +
  geom_point( size=5) +
  scale_color_manual(values = season_color_vals)+
  scale_fill_manual(values = season_color_vals) +
  theme_bw(base_size=24) +
  labs(shape = "", color="") +
  scale_x_discrete(labels=c("18 - JWC" = "JWC", "19 - PJW" = "PJW", "20 - SLEP" = "SLEP","21 - PBR" = "PBR", "22 - SFM" = "SFM", "23 - PSF" = "PSF", "24 - PNF" = "PNF", "25 - PBD" = "Canyon Mouth" )) +
  #scale_color_manual(values = cbbPalette)+
  #theme(axis.text.x = element_blank()) +
  xlab("Longitudinal Profile") +
  ylab("NO3 (mg/L)")




mainstem_plot <- ggarrange(mainstem_chla_long, mainstem_no3_long, ncol=1, nrow=2, common.legend = TRUE)
mainstem_plot
 #ggsave('output/mainstem_Longitudinal.jpg', width=13.5, height=10)
 

```

## Chl-A Timeline and Boxplots

### Main study reservoirs TIMELINE ONLY

```{r}

###CHLA TIMELINE GRAPH ####
chla_timeline <- filter(reservoir_chemistry, Year >= "2021", site_code %in% mainstem_Res_only) %>%
  filter(!is.na(ChlA))%>%
  ggplot(aes(x=Date, y=ChlA, color=site_code, group=site_code)) +
  geom_line( size=1) +
  geom_point( size=3.5) +
  scale_color_manual(values = colorsBS_site_code)+
  geom_abline(aes(slope=0, intercept=2.6), color='black', linetype="dashed") +
  geom_abline(aes(slope=0, intercept = 8), color='black', linetype="dashed") +

  labs(x = "Date", y = " Chlorophyll a (μg/L)") +
  theme_bw(base_size=24) +
  theme(legend.position = "none")+
  coord_cartesian(ylim=c(0,12))+
  facet_wrap(~Year, scales = "free_x")


plot(chla_timeline)

#ggsave('output/main_study_reservoirs_timeline_final.jpg',width=15,height=10, units ="in", dpi = 300)

```

### South Fork Reservoirs

-   Probably won't use due to low frequency of sampling

-   If sampled in future years, can be used as pre filled baseline

```{r}
###CHLA TIMELINE GRAPH ####


SF_chla_timeline <- filter(reservoir_chemistry, Year >= "2022" & site_code %in% SF_Res)%>%
  ggplot(aes(x=Date, y=ChlA, group=site_code, color=Watershed_Level)) +
  geom_path(size=1) +
  geom_point(size=3.5) +
  
  geom_textpath(data = dplyr::filter(reservoir_chemistry,Year >= "2022"& site_code %in% c("COMI","COMR","COMO")), aes(x=Date, y=ChlA, group=site_code, color=Watershed_Level, size=3, label=Site), size=4.5, fontface=2, hjust=0.2, vjust=0.0) +

  geom_textpath(data = dplyr::filter(reservoir_chemistry,Year >= "2022"& site_code %in% c("BEAV", "HORO","SFM")), aes(x=Date,y=ChlA, group=site_code, color=Watershed_Level, size=3, label=Site), size=4.5, fontface=2, hjust=0.9, vjust=0) +

  geom_textpath(data = dplyr::filter(reservoir_chemistry,Year >= "2022"& site_code %in% c("HORR", "HORI")), aes(x=Date,y=ChlA, group=site_code, color=Watershed_Level, size=3, label=Site), size=4.5, fontface=2, hjust=0.5, vjust=0)+
  
  scale_color_gradient2(name="Contributing Area Burn Intensity",low="#56B4E9", mid="#F0E442", high="#D55E00", guide="colourbar", midpoint=0.125, limits=c(0,0.24967),) +
  
  geom_abline(aes(slope=0, intercept=2.6), color='black', linetype="dashed") +
  labs(x = "Date", y = " Chlorophyll a (μg/L)") +
  theme_bw(base_size=15) +
  theme(legend.position = "none")+
  coord_cartesian(xlim = c(as.Date("2022-07-01"),as.Date("2022-10-01")),
                           ylim=c(0,14))

#+  facet_wrap(~Year, scales = "free_x")

plot(SF_chla_timeline)



###CHLA  BOXPLOT GRAPH ###
SF_chla_boxplot <- filter(reservoir_chemistry, Year >= "2022" & site_code %in% SF_Res)%>%
ggplot() + 
  geom_boxplot(aes(x=reorder(site_code,order,max),y=ChlA, middle=mean(ChlA), fill=Watershed_Level)) +
  scale_fill_gradient2(name="Watershed Burn Intensity",low="#56B4E9", mid="#F0E442", high="#D55E00", 
                        guide="colourbar", midpoint=0.125, limits=c(0,0.24967),) +
  geom_abline(aes(slope=0, intercept=2.6), color='black', linetype="dashed") +
  geom_abline(aes(slope=0, intercept=8), color='black', linetype="dashed") +
  theme_bw(base_size=28) +
  theme(legend.position = "none") +
  labs( x = "Site", y = "Chlorophyll a (μg/L)")+
  coord_cartesian(ylim=c(0,14))



plot(SF_chla_boxplot)
#ggsave('output/south_fork_reservoirs_chla_boxplot.jpg',width=15,height=10)

###COMBINING AND SAVING GRAPHS###


ggarrange(SF_chla_timeline, SF_chla_boxplot, nrow = 2, ncol = 1)
#ggsave('output/south_fork_reservoirs_chla_timeline_boxplot.jpg',width=15,height=10)
```

### mainstem

```{r}

###CHLA TIMELINE GRAPH ####
mainstem_chla <- filter(reservoir_chemistry, Year >= "2022", site_code %in% mainstem_sites )
key_mainstem_chla <- filter(mainstem_chla, site_code %in% c("JWC","SLEP","PBR","PSF","PNF","PBD"))




mainstem_chla_timeline <- filter(key_mainstem_chla, !is.na(ChlA))%>%
  ggplot() +
  geom_path(aes(x=Date, y=ChlA, group=site_code, color=factor(order)), size=1) +
  geom_point(data = key_mainstem_chla, aes(x=Date,y=ChlA, color=factor(order)), size=3.5)+
  
  # geom_textpath(data = dplyr::filter(key_mainstem_chla, site_code %in% c("JWC","SLEP" )), aes(x=Date, y=ChlA, group=site_code, color=factor(order), size=3, label=site_code), size=4.5, fontface=2, hjust=0.1, vjust=0.0) +
  # 
  # geom_textpath(data = dplyr::filter(key_mainstem_chla, site_code %in% c("PBR", "PSF")), aes(x=Date,y=ChlA, group=site_code, color=factor(order), size=3, label=site_code), size=4.5, fontface=2, hjust=0.5, vjust=0) +
  # 
  # geom_textpath(data = dplyr::filter(key_mainstem_chla, site_code %in% c("PBD", "PNF")), aes(x=Date,y=ChlA, group=site_code, color=factor(order), size=3, label=site_code), size=4.5, fontface=2, hjust=0.9, vjust=0)+
  
  #scale_color_gradient2(name="Contributing Area Burn Intensity",low="#56B4E9", mid="#F0E442", high="#D55E00", guide="colourbar", midpoint=0.225, limits=c(0,0.45576),) +
  
  geom_abline(aes(slope=0, intercept=2.6), color='black', linetype="dashed") +
  labs(x = "Date", y = " Chlorophyll a (μg/L)") +
  theme_bw(base_size=24) +
  theme(legend.position = "none")+
  scale_color_manual(values = cbbPalette)+
  coord_cartesian(xlim = c(as.Date("2022-05-01"),as.Date("2022-11-15")),
    ylim=c(0,5))

#+  facet_wrap(~Year, scales = "free_x")

plot(mainstem_chla_timeline)


###CHLA  BOXPLOT GRAPH ###
mainstem_chla_boxplot <- key_mainstem_chla%>%
  ggplot()+
  geom_boxplot(aes(x=reorder(site_code,order,max),y=ChlA, middle=mean(ChlA), fill=factor(order))) +
  #scale_fill_gradient2(name="Watershed Burn Intensity",low="#56B4E9", mid="#F0E442", high="#D55E00", guide="colourbar", midpoint=0.225, limits=c(0,0.45576),) +
  geom_abline(aes(slope=0, intercept=2.6), color='black', linetype="dashed") +
  geom_abline(aes(slope=0, intercept=8), color='black', linetype="dashed") +
  theme_bw(base_size=24) +
  theme(legend.position = "none") +
  scale_fill_manual(values = cbbPalette)+
  labs( x = "Site", y = "Chlorophyll a (μg/L)")+
  coord_cartesian(ylim=c(0,5))


plot(mainstem_chla_boxplot)
#+facet_wrap(~Year, scales = "free_x")



###COMBINING AND SAVING GRAPHS###


ggarrange(mainstem_chla_timeline, mainstem_chla_boxplot, nrow = 2, ncol = 1)
ggsave('output/chla_mainstem_timeline_boxplot.jpg',width=15,height=10)

#Longitudinal +Boxplot
ggarrange(mainstem_chla_long, mainstem_chla_boxplot, nrow = 2, ncol = 1)
#ggsave('output/chla_mainstem_longitudinal_boxplot.jpg',width=15,height=10)







```

# Annual data side by side

Figure out how to make a date object with only month and day to be able to display 2021 vs 2022 on the same timeline graph.

Chla MDL .05 mg/L

```{r}

#
study_reservoir_chla <- filter(reservoir_chemistry, site_code %in% mainstem_res_set & Year >2020)%>%
filter(!is.na(ChlA))%>%
  mutate(site_year = paste(site_code, Year, sep = " "),
         dayofyear = yday(Date)) 
         
         
chla_boxplot <- study_reservoir_chla%>%
ggplot() + 
  
  geom_boxplot(aes(x=reorder(site_year,order,max),y=ChlA, middle=mean(ChlA), fill=Watershed_Level, colour = factor(Year))) +
  scale_color_grey(start = 0.4, end = 0)+
  geom_rect(data=study_reservoir_chla, inherit.aes=FALSE, aes(xmin="LNGR 2021", xmax="LNGO 2022", ymin=-.5,
                ymax=0), color="transparent", fill="#0072B2", alpha=0.01)+
  geom_rect(data=study_reservoir_chla, inherit.aes=FALSE, aes(xmin="JOEI 2021", xmax="JOER 2022", ymin=-.5,
                ymax=0), color="transparent", fill="#56B4E9", alpha=0.01)+
  geom_rect(data=study_reservoir_chla, inherit.aes=FALSE, aes(xmin="CBRI 2021", xmax="CHD 2022", ymin=-.5,
                ymax= 0), color="transparent", fill="#E2DC62", alpha=0.01)+
  geom_rect(data=study_reservoir_chla, inherit.aes=FALSE, aes(xmin="PTRR 2021", xmax="PTRO 2022", ymin=-.5,
                ymax=0), color="transparent", fill="#E69F00", alpha=0.01)+
  geom_rect(data=study_reservoir_chla, inherit.aes=FALSE, aes(xmin="BRNR 2021", xmax="BMD 2022", ymin=-.5,
                ymax=0), color="transparent", fill="#D55E00", alpha=0.01)+
  geom_text(aes(x= 2.5, y = -.25, label= "Long Draw",  color = "black"))+
   geom_text(aes(x= 6.5, y = -.25, label= "Joe Wright",  color = "black"))+
   geom_text(aes(x= 11.5, y = -.25, label= "Chambers",  color = "black"))+
   geom_text(aes(x= 16.5, y = -.25, label= "Peterson",  color = "black"))+
   geom_text(aes(x= 20.5, y = -.25, label= "Barnes Meadow",  color = "black"))+
  scale_fill_gradient2(name="Watershed Burn Intensity",low="#56B4E9", mid="#F0E442", high="#D55E00", 
                        guide="colourbar", midpoint=0.225, limits=c(0,0.45576),) +
  geom_abline(aes(slope=0, intercept=2.6), color='black', linetype="dashed") +
  geom_abline(aes(slope=0, intercept=0.05), color='black', linetype="solid") +
  geom_abline(aes(slope=0, intercept=8), color='black', linetype="dashed") +
  annotate(geom="text", label= "MDL Limit (μg/L)", x="CBRR 2022", y=.075, vjust= -1)+
  annotate(geom="text", label= "Mesotrophic Conditions (μg/L)", x="JOEI 2021", y=2.6, vjust= -1)+
  annotate(geom="text", label= "Eutrophic Conditions (μg/L)", x="LNGR 2022", y=8, vjust= -1)+
  theme_bw(base_size=15) +
  theme(legend.position = "none")+
  #theme(axis.text.x = ifelse(study_reservoir_chla$Location == "Reservoir",element_text(face = "bold"),element_text(face = "plain"))) +
  labs( x = "Site", y = "Chlorophyll a (μg/L)")+
  coord_cartesian(ylim=c(0,12))+
  scale_x_discrete(labels = function(site_year) str_wrap(site_year, width = 2))

plot(chla_boxplot)

#ggsave("output/chla_Boxplot_main_res_2021_2022.jpg", width = 15, height =9)


```

### Res only

```{r}


chla_boxplot_Res <- filter(study_reservoir_chla, site_code %in% mainstem_Res_only )%>%
ggplot(aes(x=reorder(site_year,order,max),y=ChlA, middle=mean(ChlA), fill=Watershed_Level, colour = factor(Year))) + 
   geom_dotplot(stackgroups = FALSE, binaxis = "y", method = "histodot",stackdir = "center", position = "dodge", binwidth = .3, dotsize = 1)+
  scale_color_grey(start = 0.4, end = 0)+
  
  geom_rect( inherit.aes=FALSE, aes(xmin="LNGR 2021", xmax="LNGR 2022", ymin=-.5,
                ymax=0), color="transparent", fill="#0072B2", alpha=0.01)+
  geom_rect( inherit.aes=FALSE, aes(xmin="JOER 2021", xmax="JOER 2022", ymin=-.5,
                ymax=0), color="transparent", fill="#56B4E9", alpha=0.01)+
  geom_rect( inherit.aes=FALSE, aes(xmin="CBRR 2021", xmax="CBRR 2022", ymin=-.5,
                ymax= 0), color="transparent", fill="#E2DC62", alpha=0.01)+
  geom_rect( inherit.aes=FALSE, aes(xmin="PTRR 2021", xmax="PTRR 2022", ymin=-.5,
                ymax=0), color="transparent", fill="#E69F00", alpha=0.01)+
  geom_rect( inherit.aes=FALSE, aes(xmin=8.75, xmax=10.25, ymin=-.5,
                ymax=0), color="transparent", fill="#D55E00", alpha=0.01)+
  geom_text(aes(x= 1.5, y = -.25, label= "Long Draw",  color = "black"), size = 6)+
   geom_text(aes(x= 3.5, y = -.25, label= "Joe Wright",  color = "black"), size = 6)+
   geom_text(aes(x= 5.5, y = -.25, label= "Chambers",  color = "black"), size = 6)+
   geom_text(aes(x= 7.5, y = -.25, label= "Peterson",  color = "black"), size = 6)+
   geom_text(aes(x= 9.5, y = -.25, label= "Barnes Meadow",  color = "black"), size = 6)+
  scale_fill_gradient2(name="Watershed Burn Intensity",low="#56B4E9", mid="#F0E442", high="#D55E00", 
                        guide="colourbar", midpoint=0.225, limits=c(0,0.45576),) +
  geom_abline(aes(slope=0, intercept=2.6), color='black', linetype="dashed") +
  geom_abline(aes(slope=0, intercept=0), color='black', linetype="solid") +
  geom_abline(aes(slope=0, intercept=8), color='black', linetype="dashed") +
  #annotate(geom="text", label= "MDL Limit (μg/L)", x="CBRR 2022", y=.075, vjust= -1)+
  annotate(geom="text", label= "Mesotrophic", x=1, y=2.6, vjust= -1, size = 5)+
  annotate(geom="text", label= "Eutrophic", x= 1, y=8, vjust= -1, size = 5)+
  theme_bw(base_size=25) +
  
  #theme(axis.text.x = ifelse(study_reservoir_chla$Location == "Reservoir",element_text(face = "bold"),element_text(face = "plain"))) +
  labs( x = "Site", y = "Chlorophyll a (μg/L)")+
  coord_cartesian(ylim=c(0,12))+
  scale_x_discrete(labels = c("2021", "2022","2021", "2022","2021", "2022","2021", "2022","2021", "2022"))+
  theme(legend.position = "none",
        axis.title = element_text(face = "bold"),
         axis.text.x = element_text(face = "bold"))

plot(chla_boxplot_Res)

#ggsave("output/chla_dotplot_side_side_res_only.jpg", width = 15, height =9)

#dotplot and timeline
ggarrange(chla_timeline, chla_boxplot_Res, nrow = 2, ncol = 1)
#ggsave("output/chla_dotplot_side_side_and_timeline.jpg", width = 15, height =15)


```

## Chla timeline Res Only

```{r}
 xaxis_dates <-c("June 1st", "July 1st",  "August 1st", "Sept 1st",  "October 1st" , "November 15th")
#CHLA timeline

chla_timeline <- filter(study_reservoir_chla, site_code %in% mainstem_Res_only )%>%
  ggplot(aes(x= dayofyear, y=ChlA, group=site_year, color=site_year)) +
  geom_line(size=1) +
  geom_point(size=3.5)+
  #scale_color_gradient2(name="Contributing Area Burn Intensity",low="#56B4E9", mid="#F0E442", high="#D55E00", guide="colourbar", midpoint=0.225, limits=c(0,0.45576),) +
  geom_abline(aes(slope=0, intercept=2.6), color='black', linetype="dashed") +
  labs(x = "Date", y = " Chlorophyll a (μg/L)") +
  theme_bw(base_size=15) +
  theme(legend.position = "none")+
  coord_cartesian(ylim=c(0,12))+
  scale_x_continuous(breaks = c(150,181,215,245,275,320),labels=xaxis_dates)+
    gghighlight(site_year == "CBRR 2022"|site_year == "CBRR 2021", keep_scales = T, use_direct_label = T)

 

plot(chla_timeline)

#ggsave("output/chla_timeline_CBRR_2021_2022.jpg", width = 11, height = 9)

```

## Longitudinal Reservoir and mainstem

```{r}


#Dataframe

longitudinal_res_mainstem <- filter(reservoir_chemistry, Year == "2022"& Date != "2022-09-19")%>%
  filter ( site_code %in% c("JOEI","JOER","CBRI","CBRR","CHD", "JWC","SLEP","PBR","PSF","PNF","PBD"))%>%
  select(site_code,Date, watershed, Watershed_Level, distance_upstream_km, ChlA, NO3, PO4, K, status, location )%>%
  mutate(wk = as.character(week(Date)),
          dist_pbd = as.character(distance_upstream_km),
         type = ifelse(location %in% c("Inflow","Outflow"),"Stream",
                  ifelse(location == "Reservoir", "Reservoir", "Mainstem")))%>%
  filter(wk %in% c("22","28","32", "38", "42") )
  

longitudinal_res_mainstem$wk <- factor(longitudinal_res_mainstem$wk, levels = c("22","28","32", "38", "42"))
longitudinal_res_mainstem$type <- factor(longitudinal_res_mainstem$type, levels = c("Stream", "Reservoir", "Mainstem"))



#graphing

longitudinal_graph_res_mainstem <- longitudinal_res_mainstem%>%
  ggplot()+
  geom_point(aes(x = distance_upstream_km, y= ChlA, shape = type, color = wk), alpha = .9, size = 8)+
  geom_line(aes(x= distance_upstream_km, y = ChlA, group = wk, color =  wk), size = 4, alpha = .8)+
  geom_point(aes(x= distance_upstream_km, y= ChlA, shape = type, color = wk), alpha = .9, size = 8)+
  #gghighlight(wk == "30", use_direct_label = FALSE )+
  theme_bw(base_size = 20)+
  scale_x_reverse()+
scale_color_manual(name = "Week of", labels= c("6-01-22",
                                               "7-11-22",
                                               "8-08-22", 
                                               "9-20-22",
                                               "10-18-22"), values = c("22" = "#047E82", "28" = "#397534",  "32"= '#59B851', "38" = '#DEB907', "42" = '#FA850E' ))+
  
  scale_shape_manual(values=c(17, 16, 15))+
  scale_fill_manual( values = c("22" = "#047E82", "28" = "#397534",  "32"= '#59B851', "38" = '#DEB907', "42" = '#FA850E' ))+
 # scale_color_manual(name = "Week of", values = c("6-01" = "#047E82", "7-11" = "#397534", "7-25"='#59B851', "8-08"= '#DEB907' ))+
  labs(shape = "Site Type", color="", fill ="") +
  theme(legend.justification = c(.9,.9) , legend.position = c(.9, .9),
        #legend.box.background  = element_rect(color = "black",size = 1.5),
          axis.title = element_text(face = "bold"),
         axis.text.x = element_text(face = "bold"))+
   #theme(axis.text.x = element_blank()) +
   xlab("Distance from Canyon Mouth (km)") +
   ylab("Chlorophyll a (μg/L)")
  
plot(longitudinal_graph_res_mainstem)
#ggsave("output/longitudinal_res_main_2022.jpg", width = 20, height = 9)
  

#gghighlight graphs for presentation

  longitudinal_gghighlight_graph_res_mainstem <- longitudinal_graph_res_mainstem+
    gghighlight(wk == "42", keep_scales = T, use_direct_label = FALSE)
  
  
plot(longitudinal_gghighlight_graph_res_mainstem)
#ggsave("output/longitudinal_highlighted_res_main_2022.jpg", width = 20, height = 9)
```

### Monthly averages of \^\^

```{r}

monthly_long <- filter(reservoir_chemistry, Year == "2022")%>%
  filter ( site_code %in% c("JOEI","JOER","CBRI","CBRR","CHD", "JWC","SLEP","PBR","PSF","PNF","PBD"))%>%
  select(site_code,Date, watershed, Watershed_Level, distance_upstream_km, ChlA, NO3, PO4, K, status, location )%>%
  mutate(month = as.character(month(Date)),
          dist_pbd = distance_upstream_km,
         type = ifelse(location %in% c("Inflow","Outflow"),"Stream",
                  ifelse(location == "Reservoir", "Reservoir", "Mainstem")))%>%
  filter(!is.na(ChlA))%>%
  reframe(mean_chla = mean(ChlA),
          dist_pbd = dist_pbd,
          type = type,
          .by = c(month,site_code))%>%
  distinct()

monthly_long$month <- factor(monthly_long$month, levels = c("6","7","8", "9", "10"))


monthly_long_chla <- filter(monthly_long, month %in% c("6", "7", "8", "9", "10"))%>%
  ggplot()+
  geom_point(aes(x = dist_pbd, y= mean_chla, shape = type, color = month), alpha = .9, size = 8)+
  geom_line(aes(x= dist_pbd, y = mean_chla, group = month, color =  month), size = 4, alpha = .8)+
  geom_point(aes(x= dist_pbd, y= mean_chla, shape = type, color = month), alpha = .9, size = 8)+
  theme_bw(base_size = 20)+
  scale_x_reverse()+
scale_color_manual(name = "Month", labels= c("June",
                                               "July",
                                               "August",
                                               "September",
                                             "October"), values = c("6" = "#047E82", "7" = "#397534",  "8" = '#59B851', "9" = '#DEB907', "10" = '#FA850E' ))+

  scale_shape_manual(values=c(17, 16, 15))+
  scale_fill_manual( values = c("6" = "#047E82", "7" = "#397534",  "8"= '#59B851', "9" = '#DEB907', "10" = '#FA850E' ))+
  labs(shape = "Site Type", color="", fill ="") +
  theme(legend.justification = c(.9,.9) , legend.position = c(.9, .9),
        #legend.box.background  = element_rect(color = "black",size = 1.5),
          axis.title = element_text(face = "bold"),
         axis.text.x = element_text(face = "bold"))+
   #theme(axis.text.x = element_blank()) +
   xlab("Distance from Canyon Mouth (km)") +
   ylab("Mean Monthly Chlorophyll a (μg/L)")

plot(monthly_long_chla)

#ggsave("output/monthly_long_res_main_2022.jpg", width = 20, height = 9)
```

## Inflow/outflow ratio

```{r}



joer_cbrr_ratios <- filter(reservoir_chemistry, site_code %in% c("JOEI",
                                                                "JOER",
                                                                "CBRI",
                                                                "CBRR",
                                                                "CHD"))%>%
      dplyr::select(Date,
         site_code,
         ChlA)%>%
  drop_na()%>%
  distinct()%>%
  pivot_wider(names_from = site_code, values_from = ChlA)%>%
  mutate(jw_ratio = CBRI / JOEI,
         chambers_ratio = CHD/ CBRI, 
         Year = year(Date))
  

res_ratio_plot <- filter(joer_cbrr_ratios, Date != "2022-06-14")%>%
  ggplot()+
  geom_line(aes(x=Date, y = jw_ratio, colour ="Joe Wright"), size = 3)+
  geom_line(aes(x=Date, y = chambers_ratio, colour ="Chambers"), size = 3)+
  geom_point(aes(x=Date, y = chambers_ratio, size = CBRR, colour ="Chambers" ))+
  geom_point(aes(x=Date, y = jw_ratio, size = JOER, colour ="Joe Wright"))+
   #geom_text(aes(as.Date("2021-06-15"), 1, label= "Inflow = Outflow", vjust=-1))+
  scale_size_continuous(range = c(2,17))+
  theme_bw(base_size=25) +
  geom_abline(aes(slope=0, intercept=1), color='black', linetype="solid", label = "Inflow = Outflow")+
  labs(size = "Chl-a Levels at Reservoir")+
  scale_color_manual(name = "Reservoir", values = c("Chambers" = "#E69F00", "Joe Wright" = "#56B4E9"))+
 # theme(axis.title.y = element_text(angle = 0, vjust = .5)) +
  labs( x = "Date", y = "Ratio: Outflow Chl-a / Inflow Chl-a" )+
  facet_wrap( ~Year, scales = "free_x" )+
  
  theme(legend.justification = c(.05, .9), legend.position = c(.05, .9),legend.box = "horizontal",
          axis.title = element_text(face = "bold"),
         axis.text.x = element_text(face = "bold"))

#ggsave("output/Res_Ratio.jpg",width = 22, height = 14)
plot(res_ratio_plot)

```

## BRNR Open

This graph will highlight the changes in the chla levels downstream before and after BRNR was open

Timeline of CHD, BMD, JWC, SLEP, PBR

```{r}

barnes_effect_dataset <- filter(reservoir_chemistry, Year == "2022")%>%
  filter ( site_code %in% c("BRNR","BMD","JWC","CBRR","CHD", "JWC","SLEP","PBR"))%>%
  select(site_code,Date, watershed, Watershed_Level, distance_upstream_km, ChlA, NO3, PO4, K, status, location )%>%
  mutate(wk = week(Date),
          dist_pbd = as.character(distance_upstream_km),
         type = ifelse(location=="Outflow","Outflow",
                  ifelse(location == "Reservoir", "Reservoir", "Mainstem")))%>%
 filter(wk == 40 | wk == 42)%>%
  mutate(wk = as.character(wk),
         dist = as.character(distance_upstream_km))


barnes_effect_dataset$site_code <- factor(barnes_effect_dataset$site_code, levels = c("CBRR","CHD", "BRNR", "BMD", "JWC","SLEP","PBR"))

barnes_effect_dataset$dist <- factor(barnes_effect_dataset$dist, levels = c("87.814
", "86.049", "85.149", "84.049", "80.916","73.161","47.262"))




longitudinal_barnes_effect_graph <- barnes_effect_dataset%>%
  ggplot(aes(x= site_code, y = ChlA, group = wk, color =  wk))+
  geom_point(aes( shape = type), alpha = .9, size = 8)+
  geom_line( size = 4, alpha = .8)+
   geom_point(aes(shape = type), alpha = .9, size = 8)+
  #gghighlight(wk == "30", use_direct_label = FALSE )+
  
  theme_bw(base_size = 20)+
scale_color_manual(name = "Week of", labels= c("10-04-22", 
                                               "10-21-22"), 
                   values = c( "40" = "#397534",  "42"= '#DEB907' ))+
  
  
  scale_shape_manual(values=c(17, 16, 15))+
  scale_fill_manual( values = c("40" = "#397534",  "42"= '#DEB907' ))+
 # scale_color_manual(name = "Week of", values = c("6-01" = "#047E82", "7-11" = "#397534", "7-25"='#59B851', "8-08"= '#DEB907' ))+
  labs(shape = "Site Type", color="", fill ="") +
  theme(legend.justification = c(1,.9) , legend.position = c(1, .9),
        #legend.box.background  = element_rect(color = "black",size = 1.5),
          axis.title = element_text(face = "bold"),
         axis.text.x = element_text(face = "bold"))+
   #theme(axis.text.x = element_blank()) +
   xlab("Site Name") +
   ylab("Chlorophyll a (μg/L)")

plot(longitudinal_barnes_effect_graph)
#ggsave("output/barnes_oct_release_longitudinal.jpg", width = 12, height = 6)
```
